move readgrid classes into subcomponents module
